refactor(utils): extract gemini parsing into gemini_parse.py; utils is now a facade
Move the remaining gemini code -- parse_gemini_url, format_gemini_url,
validate_gemini_url_components, parse_gemini_response and the
process_gemini_response result pipeline (with its _process_* helpers) -- into
gemini_parse.py, which delegates gemtext rendering and MIME handling to the
gemtext and mime modules.

utils.py is now a thin re-export facade: the ~1700-line module is replaced by
five focused modules (helpers, mime, gemtext, gopher_parse, gemini_parse), each
re-exported so every `from gopher_mcp.utils import X` keeps working unchanged.
No behavior change.
